[Refactor] encapsulate the algorithm for finding the drop position
