+++ src/editor/contrib/dragAndDropExtension/dragAndDropExtension.ts
@@ -4,6 +4,7 @@ import { EditorExtension, IEditorExtension } from "src/editor/common/editorExten
 import { ProseEditorView } from "src/editor/common/proseMirror";
 import { EditorExtensionIDs } from "src/editor/contrib/builtInExtensionList";
 import { IEditorWidget } from "src/editor/editorWidget";
+import { getDropExactPosition } from "src/editor/common/cursorDrop";
 
 export interface IEditorDragAndDropExtension extends IEditorExtension {
     readonly id: EditorExtensionIDs.DragAndDrop;
@@ -150,46 +151,9 @@ export class EditorDragAndDropExtension extends EditorExtension implements IEdit
         if (!view.editable) {
             return;
         }
-    
-        /**
-         * Handle the case when the mouse is inside the editor, we try to find
-         * the node at the exact mouse position (clientX and clientY).
-         */
-        const position = view.posAtCoords({ left: event.clientX, top: event.clientY });
-        if (position && position.inside !== -1) {
-            let target: number = position.pos;
-            if (view.dragging) {
-                const point = dropPoint(view.state.doc, target, view.dragging.slice);
-                if (point !== null) {
-                    target = point;
-                }
-            }
-            this.__renderCursor(target, view);
-        }
-        /**
-         * Reach here might be either:
-         *  1. `position` is null, means the mouse is dragging outside the editor.
-         *  2. `position.inside = -1`, which means it is the node at the top level.
-         * Either case, we render the cursor only in block-level.
-         */
-        else {
-            const mouseY = event.clientY;
-            const viewRect = view.dom.getBoundingClientRect();
-            const position = view.posAtCoords({ left: viewRect.left, top: mouseY });
-            
-            // a node found at the mouse-y axis.
-            if (position) {
-                const target = position.inside === -1 ? position.pos : position.inside;
-                this.__renderCursor(target, view);
-            } 
-            // node not found at y axis. Must be either above/below the entire editor.
-            else {
-                const docSize = view.state.doc.content.size;
-                const target = mouseY > docSize ? docSize : 0;
-                this.__renderCursor(target, view);
-            }
-        }
-
+        const position = getDropExactPosition(view, event);
+        this.__renderCursor(position, view);
+        
         event.preventDefault();
     }
